refactor eventset so that server side can depend on proto module to do event set related stuff
