+++ stream/clients/java/stream/pom.xml
@@ -30,11 +30,6 @@
   <dependencies>
     <dependency>
       <groupId>org.apache.bookkeeper</groupId>
-      <artifactId>circe-checksum</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.bookkeeper</groupId>
       <artifactId>stream-storage-java-client-base</artifactId>
       <version>${project.parent.version}</version>
     </dependency>
