+++ stream/clients/java/stream/src/test/java/org/apache/bookkeeper/clients/impl/stream/event/ReadEventsImplTest.java
@@ -26,7 +26,8 @@ import io.netty.buffer.ByteBuf;
 import java.io.IOException;
 import java.util.concurrent.ThreadLocalRandom;
 import org.apache.bookkeeper.api.stream.ReadEvent;
-import org.apache.bookkeeper.clients.impl.stream.event.EventSet.Writer;
+import org.apache.bookkeeper.stream.protocol.EventSet;
+import org.apache.bookkeeper.stream.protocol.EventSet.Writer;
 import org.apache.bookkeeper.common.coder.StringUtf8Coder;
 import org.apache.bookkeeper.common.coder.VarIntCoder;
 import org.apache.bookkeeper.stream.proto.CompressionCodecType;
