+++ stream/clients/java/stream/src/test/java/org/apache/bookkeeper/clients/impl/stream/event/TestEventSet.java
@@ -30,8 +30,9 @@ import java.util.List;
 import java.util.Random;
 import java.util.concurrent.atomic.AtomicReference;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.bookkeeper.clients.impl.stream.event.EventSet.Reader;
-import org.apache.bookkeeper.clients.impl.stream.event.EventSet.Writer;
+import org.apache.bookkeeper.stream.protocol.EventSet;
+import org.apache.bookkeeper.stream.protocol.EventSet.Reader;
+import org.apache.bookkeeper.stream.protocol.EventSet.Writer;
 import org.apache.bookkeeper.common.coder.Coder;
 import org.apache.bookkeeper.common.coder.StringUtf8Coder;
 import org.apache.bookkeeper.common.coder.VarIntCoder;
