+++ stream/clients/java/stream/src/test/java/org/apache/bookkeeper/clients/impl/stream/range/StreamRangeEventReaderImplTest.java
@@ -32,7 +32,6 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import org.apache.bookkeeper.api.stream.ReadEvent;
 import org.apache.bookkeeper.api.stream.ReadEvents;
-import org.apache.bookkeeper.clients.impl.stream.event.EventSet;
 import org.apache.bookkeeper.clients.impl.stream.event.ReadEventsImpl;
 import org.apache.bookkeeper.clients.impl.stream.event.ReadEventsImpl.Recycler;
 import org.apache.bookkeeper.common.coder.StringUtf8Coder;
@@ -42,6 +41,7 @@ import org.apache.bookkeeper.stream.proto.CompressionCodecType;
 import org.apache.bookkeeper.stream.proto.EventSetType;
 import org.apache.bookkeeper.stream.proto.RangePosition;
 import org.apache.bookkeeper.stream.proto.RangeProperties;
+import org.apache.bookkeeper.stream.protocol.EventSet;
 import org.apache.bookkeeper.stream.protocol.RangeId;
 import org.junit.After;
 import org.junit.Before;
