+++ stream/proto/pom.xml
@@ -48,6 +48,11 @@
       <version>${project.version}</version>
       <scope>test</scope>
     </dependency>
+    <dependency>
+          <groupId>org.apache.bookkeeper</groupId>
+          <artifactId>circe-checksum</artifactId>
+        <version>${project.version}</version>
+    </dependency>
   </dependencies>
 
   <build>
