+++ stream/proto/src/main/java/org/apache/bookkeeper/stream/protocol/EventSet.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.bookkeeper.clients.impl.stream.event;
+package org.apache.bookkeeper.stream.protocol;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkNotNull;
